# perf(@angular-devkit/build-angular): reduce build times for apps with  a large number of components when utilizing esbuild-based builders